Added Multiwfn_ELF_Heatmap.py
A little python script that can be used to command-line the creation of ELF plots using a plane defined by 3 atoms and offset by a provided distance.  Wrote this to stop having to manually calculate plane equations based on points and desired offsets and then added Multiwfn silent usage.

This does require that you a) have Multiwfn installed and in your system path, and b) that in the settings.ini file, you have "isilent=1" (not the default setting in there).